refactor: split asset protocol module
